feat: a billing page for failed charges
