+++ docs/api/index.md
@@ -1,1 +1,5 @@
 Nothing here yet! To get started and involved, [check out our GitHub page](https://github.com/funmusicplace/mirlo).
+
+If you want to use Mirlo's API, you'll need an API KEY. This can be set locally by editing the client table in the database, or if you want to work against the live API, you have to request a token by e-mailing hi@mirlo.space
+
+Documentation for the API can be found at https://mirlo.space/docs
