+++ emails/charge-failure/subject.pug
@@ -1,0 +1,1 @@
+= `Charge failure: Your payment method could not be processed`
