+++ src/routers/v1/artists/{id}/purchaseCatalogue.ts
@@ -3,7 +3,7 @@ import { NextFunction, Request, Response } from "express";
 import { userLoggedInWithoutRedirect } from "../../../../auth/passport";
 import prisma from "@mirlo/prisma";
 
-import { createStripeCheckoutSessionForCatalogue } from "../../../../utils/stripe";
+import { createStripeCheckoutSessionForCatalogue } from "../../../../utils/stripe/sessions";
 import { subscribeUserToArtist } from "../../../../utils/artist";
 import { AppError } from "../../../../utils/error";
 import { determinePrice } from "../../../../utils/purchasing";
