+++ src/routers/v1/checkout/status.ts
@@ -1,5 +1,4 @@
 import { NextFunction, Request, Response } from "express";
-import prisma from "@mirlo/prisma";
 
 import stripe from "../../../utils/stripe";
